refactor : 패키지 분리, 출력 메세지 분리
