--- src/main/java/calculator/OutputView.java
@@ -1,9 +1,0 @@
-package calculator;
-
-public class OutputView {
-    private final String outputMessage = "결과 : ";
-
-    public void printResult(Long resultNumber) {
-        System.out.println(outputMessage + resultNumber);
-    }
-}
