+++ src/main/java/calculator/exception/ErrorMessage.java
@@ -1,0 +1,6 @@
+package calculator.exception;
+
+public class ErrorMessage {
+    public static final String NOT_NEGATIVE_NUM = "양수만 입력해 주세요.";
+    public static final String NOT_CONVERTED_NUM = "숫자로 변환할 수 없습니다.";
+}
