+++ src/main/java/calculator/io/InputView.java
@@ -1,12 +1,12 @@
-package calculator;
+package calculator.io;
+
 import camp.nextstep.edu.missionutils.Console;
 
 public class InputView {
-    private static final String START_CALCULATE = "덧셈할 문자열을 입력해 주세요.";
     private final String input;
 
     public InputView() {
-        System.out.println(START_CALCULATE);
+        System.out.println(consoleMessage.START_CALCULATE);
         input = Console.readLine();
     }
 
@@ -14,7 +14,7 @@ public class InputView {
         return input;
     }
 
-    public void closedConsole(){
+    public void closedConsole() {
         Console.close();
     }
 }
