+++ src/main/java/calculator/io/OutputView.java
@@ -1,0 +1,8 @@
+package calculator.io;
+
+public class OutputView {
+
+    public void printResult(Long resultNumber) {
+        System.out.println(consoleMessage.OUTPUT_MESSAGE + resultNumber);
+    }
+}
