+++ src/main/java/calculator/io/consoleMessage.java
@@ -1,0 +1,6 @@
+package calculator.io;
+
+public class consoleMessage {
+    public static final String START_CALCULATE = "덧셈할 문자열을 입력해 주세요.";
+    public static final String OUTPUT_MESSAGE = "결과 : ";
+}
